Update theme-localization.md
I think it should be highlighted that when using Polylang the 'Generating Language Files' and 'Loading Language Files' section can be skipped. I spent a couple of hours trying to get that all working on WSL and it turns out the solution with Polylang and the 'BenjaminMedia/wp-polylang-theme-strings' plugin saves all that messing around.